:art: Improving some definitions about boarding looping, and removing linter warnings
