+++ .rubocop.yml
@@ -7,7 +7,7 @@ AllCops:
   DisplayCopNames: true
 
 Layout/LineLength:
-  Max: 125
+  Max: 120
 Metrics/MethodLength:
   Max: 20
 Metrics/AbcSize:
@@ -17,8 +17,6 @@ Metrics/ClassLength:
 Metrics/BlockLength:
   ExcludedMethods: ['describe']
   Max: 30
-Metrics/ModuleLength:
-  Max: 500
 
 
 Style/Documentation:
@@ -33,8 +31,6 @@ Style/DefWithParentheses:
   Enabled: false
 Style/FrozenStringLiteralComment:
   EnforcedStyle: never
-Style/CaseEquality:
-  Enabled: false
 
 Layout/HashAlignment:
   EnforcedColonStyle: key
