+++ .stickler.yml
@@ -11,3 +11,6 @@ files:
     - "Rakefile"
     - "node_modules/**/*"
     
+# PLEASE DO NOT enable auto fixing options
+# if you need extra support from you linter - do it in your local env as described in README for this config
+# find full documentation here: https://stickler-ci.com/docs
